Refactor model scoring and remove unused files
